Changed to use Sapper server routes
