+++ api/package.json
@@ -13,7 +13,7 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@types/dotenv": "^8.2.0",
+    "dotenv": "^8.2.0",
     "apollo-boost": "^0.3.1",
     "apollo-cache-inmemory": "^1.6.3",
     "apollo-client": "^2.6.4",
