+++ ui/package.json
@@ -19,12 +19,12 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "@types/dotenv": "^8.2.0",
     "compression": "^1.7.1",
     "polka": "next",
     "sirv": "^0.4.0"
   },
   "devDependencies": {
+    "dotenv": "^8.2.0",
     "@babel/core": "^7.0.0",
     "@babel/plugin-syntax-dynamic-import": "^7.0.0",
     "@babel/plugin-transform-runtime": "^7.0.0",
