+++ ui/src/routes/categories/index.svelte
@@ -1,13 +1,9 @@
 <script context="module">
 	export async function preload({ params, query }) {
-    console.log("retrieving");
 		const res = await this.fetch(`categories.json`);
-    console.log("retrieved");
-
 		if (res.status === 200) {
       const categories = await res.json();
-      console.log(categories);
-			return { categories };
+      return { categories };
 		}
 
 		this.error(500, 'Could not load users');
@@ -31,18 +27,11 @@
 
 <h1>Categories</h1>
 
-{#await $categories}
-  <p>Loading...</p>
-{:then result}
+<ul>
+  {#each categories.Category as { name }}
+    <li>
+      <a rel="prefetch" href="categories/{name}">{name}</a>
+    </li>
+  {/each}
+</ul>
 
-  <ul>
-    {#each result.data.Category as { name }}
-      <li>
-        <a rel="prefetch" href="categories/{name}">{name}</a>
-      </li>
-    {/each}
-  </ul>
-
-{:catch error}
-  <p>Error: {error}</p>
-{/await}
