+++ ui/src/routes/reviews/index.json.js
@@ -22,14 +22,12 @@ query getUsers {
 `;
 
 export async function get(req, res, next) {
-    console.log("query users");
     const users = await client.query({
         query: GET_USERS
     });
-    console.log(devalue(users.data));
 	if (users !== null) {
-        res.setHeader('Content-Type', 'text/plain; charset=UTF-8');
-		res.end(devalue(users.data));
+		res.setHeader('Content-Type', 'application/json');
+		res.end(JSON.stringify(users.data));
 	} else {
 		next();
 	}
